Backend/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.3
pinecone-client==3.2.2 
trafilatura
langdetect
deep-translator
pytest
pypdf                
python-docx           
unstructured[local-inference] 
docx2txt
streamlit